rishicreateprotocol-main/pyproject.toml
[tool.poetry]
name = "rishi-api"
version = "0.1.0"
description = "This is the rishi API"
authors = ["Dipankar Sarkar <me@dipankar.name>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-smorest = "^0.44.0"
dramatiq = {extras = ["redis"], version = "^1.17.0"}
requests = "^2.32.3"
python-dotenv = "^1.0.1"
replicate = "^0.31.0"
marshmallow = "^3.21.3"
boto3 = "^1.34.159"
flask-melodramatiq = "^1.0.1"
fastapi = "^0.114.0"
sqlalchemy = "^2.0.34"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pydantic = "^2.9.0"
uvicorn = "^0.30.6"
python-multipart = "^0.0.9"
click = "^8.1.7"
bcrypt = "3.2.2"
alembic = "^1.13.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
